+++ radical_translations/core/views.py
@@ -132,6 +132,11 @@ class ResourceViewSet(DocumentViewSet):
             "enabled": True,
             "options": ES_FACET_OPTIONS,
         },
+        "radical_date": {
+            "field": "has_date_radical",
+            "enabled": True,
+            "options": ES_FACET_OPTIONS,
+        },
         "meta": {"field": "meta", "enabled": True, "options": ES_FACET_OPTIONS},
     }
 
@@ -154,6 +159,7 @@ class ResourceViewSet(DocumentViewSet):
         "subject": "subjects.label.raw",
         "event": "events.title.raw",
         "event_place": "events.place.address.raw",
+        "radical_date": "has_date_radical",
         "meta": "meta",
     }